+++ src/EUTelM26Timing.cc
@@ -1,4 +1,4 @@
-// Version: $Id: EUTelM26Timing.cc 2285 2013-01-18 13:46:44Z hperrey $
+// Version: $Id: EUTelM26Timing.cc 2525 2013-04-03 07:00:59Z hamnett $
 // Author:  Georg Troska <mailto: georg.troska@uni-dortmund.de>
 
 
@@ -60,7 +60,7 @@ void EUTelM26Timing::print(std::ostream& os) const {
   }
   os << endl;
   int width = 20;
-  os << setw(width) << setiosflags(ios::left) << "Type: "     << _type << endl
+  os << setw(width) << setiosflags(ios::left) << "Type: "     << static_cast< int >(_type) << endl
      << setw(width) << "Elements: " << _noOfElements << endl
      << setw(width) << "TLU: "  << _tluCounter << endl
      << setw(width) << "Sensor: "  << _sensorID << endl;
